# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/create": {
            "post": {
                "description": "Creates a new entity in God World. Chaos may interfere.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entities"
                ],
                "summary": "Create an entity",
                "parameters": [
                    {
                        "description": "Entity info",
                        "name": "entity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/destroy/{name}": {
            "delete": {
                "description": "Removes an entity by name. Chaos may interfere.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entities"
                ],
                "summary": "Destroy an entity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Entity Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/dump": {
            "get": {
                "description": "Shows all entities currently in the world",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "world"
                ],
                "summary": "Display world Entities",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/god.Entity"
                            }
                        }
                    }
                }
            }
        },
        "/dump/world": {
            "get": {
                "description": "Shows everything about the current world",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "world"
                ],
                "summary": "Display world state",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.DumpResponse"
                        }
                    }
                }
            }
        },
        "/get/{name}": {
            "get": {
                "description": "Retrieves an entity by name, may include chaos message if chaos interfered",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entities"
                ],
                "summary": "Get an entity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Entity Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GetEntityResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.CreateRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "properties": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "api.DumpResponse": {
            "type": "object",
            "properties": {
                "entities": {
                    "description": "map of entity name → properties",
                    "type": "object",
                    "additionalProperties": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        }
                    }
                },
                "events": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "api.GetEntityResponse": {
            "type": "object",
            "properties": {
                "chaos_msg": {
                    "type": "string"
                },
                "entity": {
                    "$ref": "#/definitions/god.Entity"
                }
            }
        },
        "god.Entity": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "properties": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        }
    }
}